Extract issue metadata risky policy module
